Unwillingly training to be an assassin like Anne Parillaud in La Femme Nikita (the original Luc Besson version, of course, and not the mind-wrenchingly diabolical John Badham remake, or the TV series where suddenly she was "arrested for a crime she did not commit," like shooting a cop point-blank in the face when hopped up on drugs during a botched raid on a chemist's' is easy to pass off as a fit-up), Matthew switched courses halfway through the year and went on to pass glowingly in lawn maintenance.
On March 5th 1992 he took two bullets meant for the Pope due to a mix-up
at the post office, and, ambushed by the gang of international bullet
smugglers intending to hijack the parcel, was forced to clip their
bonces with a hoe.

Disowned by his local Hypervalue for no particular reason, Matthew
became a Drifter, appearing on their second album and successfully
lobbying against the proposed "nude cellist" look. Weeks later he had
moved on again, this time heading to Omagh, N Ireland, in search of
Gainful Employment, a 500-1 steeplechaser that, by a needlessly
complicated sting, had won decisively before vanishing with an estimated
two million pounds of bets placed in a variety of disguises. After weeks
of careful detective work, he cornered the horse, but let it go on a
whim after a large bribe.

Now a county-level swimming champion, Matthew divides his time equally
between charitable works and writing to Lambert Simnel. The joke's on
him, however, as not only was Simnel executed in the 15th century, but
the old woman now at that address is steaming off the return stamps.

Matthew would one day like to work with Laurence Marks and Maurice Gran,
preferably in a format that involves unmercifully beating Laurence Marks
and Maurice Gran.