Emit "transitioned" not "changed" for new frames
Previously, "changed" would get emitted when switching
backgrounds AND when switching slides in an animated
background.  The two actions are conceptually different,
so this commit splits the signal into two signals. This
will allow us to add a cross fade effect when switching
backgrounds (and not add the cross fade effect when
switching slides that have their own transition effect)

svn path=/trunk/; revision=5350