Move part of set_pixmap_as_root to set_root_pixmap_id
The meatiest part of set_pixmap_as_root takes the passed
in pixmap and stores it on the root window in the
ESETROOT_PMAP_ID and _XROOTPMAP_ID properties.  That
functionality stands on its own, and should be factored
out so it can get reused later when adding crossfade
transitions on background changes.

svn path=/trunk/; revision=5352