Add gnome_bg_set_pixmap_as_root_with_crossfade
This implements a crossfade transition when switching
backgrounds.  To actually get the fade, though, requires
changes to gnome-settings-daemon to use this function
instead of gnome_bg_set_pixmap_as_root.

We also expose gnome_bg_get_pixmap_from_root which will
be useful for eel when it gets changed to do fading.

svn path=/trunk/; revision=5355