bgo#556050 - Make GnomeRRLabeler's labels explicitly black
The labels are always drawn against a light pastel background.  When using an
inverse-colored theme (with text nearly white, backgrounds nearly black), we
want GnomeRRLabeler's labels to still be black.  Otherwise, the color from
the theme will be hard to read against the light pastel background in the
per-monitor labels.

Signed-off-by: Federico Mena Quintero <federico@novell.com>