Only select for RRScreenChangeNotifyMask, not for the more detailed RANDR 1.2 events
We don't care about property changes on outputs, as we only handle the
EDID_DATA property.  That property isn't likely to change on an
output, anyway.

Then, we don't do anything special when an individual output or CRTC
changes.  We just care if the screen layout changed in general; in
that case, we simply re-probe the outputs and call the caller's
callback.  The X server will emit RRScreenChangeNotify in conjunction
with the more detailed events, anyway, so we are guaranteed to know
when things change.

Signed-off-by: Federico Mena Quintero <federico@novell.com>