Don't ask for RANDR events if the caller doesn't specify a callback
If the caller of GnomeRR specifies a NULL callback, it means that the
caller is not interested in RANDR events.  In that case, it is not
necessary to actually ask the X server for those events, or to register
RANDR's events with GDK.

This prevents spurious events from being generated by gnome-settings-daemon
and gnome-control-center when they create a temporary GnomeRRScreen to save
the current configuration for backup purposes.

Signed-off-by: Federico Mena Quintero <federico@novell.com>