Accumulate errors in the recursive real_assign_crtcs()
When we need to debug why a CRTC assignment didn't succeed, we need the full
log of what the algorithm tried to do.  This accumulates the errors that
happen along the process inside real_assign_crtcs().  Before we could
silently leak the innermost errors, and we were ignoring them in the end.

Signed-off-by: Federico Mena Quintero <federico@novell.com>