[pixmaps] Remove old icons installed in /usr/share/pixmaps
We kept them in GNOME 2 for compatibility because /usr/share/pixmaps was
still in use at the beginning of GNOME 2. But for GNOME 3, we're now
completely in the icon theme world, and we don't need those anymore.

If an icon there was really important, please ask for it to be included
in gnome-icon-theme.