bgo#621046 - [randr] Force a refresh of timestamps so other clients won't think this is a server-generated event
'Detect displays' in the control center's display capplet works by calling
gnome_rr_screen_refresh().  However, this causes gnome-settings-daemon's
RANDR plug-in to get an event, which makes it think that it should reconfigure
the displays.  Here, we grab the X server and do a no-op XRRSetCrtcConfig() so
that when we ungrab, the X server will send out an RRNotify event with an
updated timestamp.  This will cause g-s-d not to reconfigure the displays, thus
making 'Detect displays' really be 'detect only', not 'detect and reconfigure'.

Signed-off-by: Federico Mena Quintero <federico@novell.com>