gnome_rr_screen_new() now creates a singleton for the specified GdkScreen
Gnome-settings-daemon has now more than plugin that requires a GnomeRRScreen.  Having two
instances of a GnomeRRScreen would mean extra roundtrips to the X server when RANDR events
come in.  Also, it is better if GObject signals can be emitted in order from a single
GnomeRRScreen instance.

So, we now create a single GnomeRRScreen instance per GdkScreen, and return that.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>