gnome-bg: publish the 'representative color'
Publish a list of 'representative colours' as a property on the root
window _GNOME_BACKGROUND_REPRESENTATIVE_COLORS.  This is alongside the
normal properties pointing at the pixmap of the background image.

Right now, this is just implemented as an average using the existing
pixbuf_average_value() function.  Strictly speaking, the property is a
list of colours and in the future we may set more than one.

https://bugzilla.gnome.org/show_bug.cgi?id=671750