gnome-bg: Fix crash when system clock is behind start date
The 1-slide background included in Fedora 17 causes a crash
in gnome-bg when the system clock is set behind the date specified
in the slideshow xml file.

This was due to slide->duration being set to UINT_MAX (to indicate a
single-slide show) but show->total_duration not being updated to take
account of this. The calculation in get_current_slide() was therefore
confused, and triggered a g_assert_not_reached().

https://bugzilla.gnome.org/show_bug.cgi?id=673551