gnome-xkb-info: Handle duplicated layouts
When merging entries from another .xml file (i.e. when
show-all-sources is active) there might be layout entries which were
already added before. In that case we don't want to create new entries
and replace the old ones because layout variants which were already
added depend on the existence of the old main layout.

https://bugzilla.gnome.org/show_bug.cgi?id=692219

https://bugzilla.gnome.org/show_bug.cgi?id=694149