gnome-xkb-info: Don't store duplicated layouts in locale tables
We can't store duplicated layouts anywhere since those get destroyed
when the 'layout' element ends and we'd end up with invalid entries
potentially leading to crashes when used.

Just doing a g_hash_table_remove() on the layouts_by_* tables was a
thinko and could never work since those only contain other tables as
values and it's in this second level tables that the layouts are
stored.

https://bugzilla.gnome.org/show_bug.cgi?id=695350