GnomeRRConfig: try again when failing application because of access denied
If we get an error due to stale configuration information, try
again a second time after refreshing the screen. This can happen
when mutter is using the X11 backend, as that causes multiple events,
so a single ApplyConfiguration() can increase the serial multiple
times, and there is a race between when mutter gets the event
(increasing the serial) and when we get it, asking for the new
configuration.

https://bugzilla.gnome.org/show_bug.cgi?id=705510