GnomeRR: Do not assume outputs with the id of 'default' are built-in
We are using this in all kinds of places to work out if a display is embedded
(like a laptop display) and cannot be removed. Some of the binary drivers do
not set a sensible ID, but it's not more sane to assume these default values are
either for internal or external panels. Assuming the former breaks VMVGA in
virtual machines. We should not return TRUE when there is no heuristic to use.

Resolves: https://bugzilla.gnome.org/show_bug.cgi?id=706738