gnome-xkb-info: Remove xkbfile dependency
This code was originally from gnome-settings-daemon but got moved here
so that we could be sure that all GnomeXkbInfo users were using the
same xkb rules file by getting it from the X server.

As we move towards wayland though, we find ourselves in a position to
specify how and where xkb data resides so we can drop this
dependency. Instead, for now, we'll just hardcode the path to the xkb
data from the xkeyboard-config XKB_DATA compile time variable.

https://bugzilla.gnome.org/show_bug.cgi?id=719942