gnome-rr: Initialize GnomeRROutputInfo:rotation to GNOME_RR_ROTATION_0
Otherwise, the GnomeRROutputInfos representing connected but inactive
outputs (ie. without an assigned CRTC) had their rotations set to
GNOME_RR_ROTATION_NEXT, which is not a valid value. This leads to
gnome_rr_config_applicable returning FALSE, because it thinks that one
of the outputs require a rotation that is not supported by any of the
available CRTCs.

https://bugzilla.gnome.org/show_bug.cgi?id=727023