thumbnail: Fix a code path where the mutex is never unlocked
Fix one unusual code path where the GnomeDesktopThumbnailFactory’s mutex
is not unlocked before the function returns. As well as the usual
deadlock issues, this would lead to an abort on shutdown with recent
GLib changes, as g_mutex_clear() would be called on the locked mutex,
which GLib now (correctly) complains about.

https://bugzilla.gnome.org/show_bug.cgi?id=740666