Drop libgsystem submodule, stop using O_NOATIME
The entire rationale for adding this _noatime function was fairly
bogus.

See https://bugzilla.gnome.org/show_bug.cgi?id=680326

But the Linux kernel doesn't permit non-root users to use it for
root-owned files - which is the default for real world GNOME
installations.  We end up doing twice the syscall traffic with no
benefit at all.

The patch does work in jhbuild, but honestly the relatime handling is
generally OK for this, and where it's not, change your system to mount
O_NOATIME.

https://bugzilla.gnome.org/show_bug.cgi?id=708453