thumbnail factory: Disconnect signal handlers in finalize
We are seeing crashes in Fedora that point at the settings signal
handlers getting run after the thumbnail factory is finalized.
Explicitly disconnecting the handlers in finalize is the right
thing to do, anyway.

While we are at it, replace some of the cleanup code in finalize
with g_clear_pointer and g_clear_object, as suggested by Colin.

https://bugzilla.gnome.org/show_bug.cgi?id=761049