thumbnail: Demote a log message from WARNING to DEBUG
Failure to create a GdkPixbufLoader for a specific MIME type doesn't
necessarily indicate an error. It is possible that the fallback code
would still be able to parse the image data. For example, Canon CR2 RAW
files have the MIME type image/x-canon-cr2. While we don't have a
loader for that specific MIME type, the TIFF loader can still parse the
data.

In case the fallback code failed to parse the image data, we get a
WARNING anyway:
  GnomeDesktop-WARNING **: Error creating thumbnail for ...

Having a log message to indicate that we are using the fallback code is
useful for debugging, but there is no need for the WARNING. It can be
extra noise and needlessly interferes with things like
G_DEBUG=fatal-warnings.

https://bugzilla.gnome.org/show_bug.cgi?id=762504