thumbnail: Split off running the script
Move most of the script command generation to a separate file, making
the function return a GBytes from a successful thumbnailer run, so as to
avoid having to clean up temporary files from the thumbnailer run.

Note that it changes a few subtle things which shouldn't be a problem in
practice, but, as a corner case, might have been used by applications:

- Thumbnailers must output PNG images. pixbuf_new_from_bytes() could
have been made more complicated to handle all images, and then we would
restrict the thumbnailer output format separately, but it makes no sense
to write complicated code to remove it in the next commit.
- URIs which have no backing path are not supported. This will likely
cause problems for thumbnailing remote shares on OSes which lack
gvfsd-fuse. Support could be re-added in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=774497