tests: Adjust test conditions to ensure assumptions
In order to test that date and time are separated by the correct
UTF-8 character on UTF-8 locales, the clock needs to show the date
in the first place.

In order to correctly test that the returned string doesn't contain
the non-UTF-8 fallback of two spaces, we need to make sure that the
date itself doesn't inject any extra spaces itself ("Aug  1").

https://gitlab.gnome.org/GNOME/gnome-desktop/issues/8