gnome-language: Fix translated language names not working when LANGUAGE env variable used
GNU gettext ignores the locale set by setlocale() when the LANGUAGE environment
variable is used. This means the code to get translated langauges only returns
the local language when LANGUAGE is set. Remove this variable around the call.

Note that changing an environment variable inside a program is risky, but the
existing code is already not thread safe (since gettext is not).