tests: Guard against default changes
The wall-clock weekday test assumes that only the time is shown when
the weekday setting is disabled. This is only true while the clock
doesn't include the date by default, which is may happen as proposed
in gsettings-desktop-schemas!2. But even if the default isn't changed
it is still safer to spell out the assumptions, so explicitly disable
the date.

https://gitlab.gnome.org/GNOME/gnome-desktop/merge_requests/18