Meson: fix preprocessor directives.
When using `#ifdef` in your code (vs `#if`), the C preprocessor doesn't
check the value of the macro, only whether it is defined at all.

By using `conf.set10()`, the macros were defined, whether the
boolean values in the meson build file were false or not. So, to fix
this, you either have to start using `#if`, or you make sure you use
`conf.set()` instead.

This fixes the flatpak build of GNOME Contacts (which turns off udev
support).