Go back to a filename parameter for GnomeBGSlideShow
I think GFile works well for the property, but not for the new function.
There's not much value in forcing API users to create a GFile
themselves, and if you already have a GFile it's trivial to get the
filename from it. Also, we forgot the (transfer none) annotation on the
parameter.