desktop-docs: Fix Galician translation
Warning: Could not merge gl translation for msgid:
You must give prominent notice with each copy of the work that the Library is used in it and that the Library and its use are covered by this License. You must supply a copy of this License. If the work during execution displays copyright notices, you must include the copyright notice for the Library among them, as well as a reference directing the user to the copy of this License. Also, you must do one of these things: <_:orderedlist-1/>

Tags aren't supposed to be translated...

Fixes: f95fe3ff2681605e3815121f23c5a96e6eec45f5