Revert "tests/wall*: Do not fail if some of the locales is missing"
Running all the tests isn't optional, and could lead to tests being
skipped if there's any sort of transient problem with the CI setup.

Seeing as those tests test things that could easily be broken by someone
who doesn't run those languages, we need the tests to be run.

We'll keep the locale setting checks though, to fail more gracefully
when locales are not installed correctly.

This reverts commit 9ae847ec875efe443d51d584fa6c61966c7c3c7c.