thumbnail: Mount /etc/alternatives if it exists
On distributions that use dpkg's update-alternatives(8), Red Hat's
reimplementation or some other fork or reimplementation of the same
concept, /etc/alternatives is conceptually a configuration layer over
part of /usr. On those distributions, we cannot rely on /usr to be
self-contained, but the combination of /usr and /etc/alternatives
probably *is* self-contained.

Resolves: https://gitlab.gnome.org/GNOME/gnome-desktop/-/issues/92
Bug-Debian: https://bugs.debian.org/943434
Signed-off-by: Simon McVittie <smcv@debian.org>