desktop-docs: Fix Catalan translation
Warning: Could not merge ca translation for msgid:
For an executable, the required form of the <quote>work that uses the Library</quote> must include any data and utility programs needed for reproducing the executable from it. However, as a special exception, the materials to be distributed need not include anything that is normally distributed (in either source or binary form) with the major components (compiler, kernel, and so on) of the operating system on which the executable runs, unless that component itself accompanies the executable.

Tags aren't supposed to be translated...

Fixes: c6e9e5c9915b0cc3a513ad4014562da94322f998


(cherry picked from commit 680d80f8480f82e09315485e226e255c77f59aa0)