thumbnail: Fix thumbnailing of CBZ with UTF-8 filenames
Fix thumbnailing of CBZ comics with UTF-8 filenames, as libarchive will
not be able to parse UTF-8 filenames if a UTF-8 locale is not used.

This assumes that the C.UTF-8 locale is available on the host system,
which shouldn't be a problem on systems able to run GNOME in the first
place.

See https://github.com/libarchive/libarchive/issues/587
and https://github.com/flatpak/flatpak/pull/4138