Add libgnome-desktop-4
For GNOME components using GTK 4, we provide a base library that does
not depend on GTK, libgnome-desktop-4, with the bits that do not use GDK
or GTK.

This library has a new soname and version, as well as a new pkg-config
file and introspection data files.

We want to keep libgnome-desktop-3 as a legacy library for GNOME
components (and some applications) that still use GTK 3; for this
reason, libgnome-desktop-3 will not change.