Port gnome-bg and gnome-rr to GTK4
In order to port the GnomeBG and GnomeRR APIs to GTK4 we need to copy
the files into their own subdirectories, as we want to keep the older
GTK3-based implementations available for the legacy
libgnome-desktop-3.0. It also makes it easier for us to spin off these
libraries into their own projects, if we decide to do so.