Use fedora:latest for CI image
Previously, we used fedora:rawhide for gnome-desktop CI, but stopped
because gitlab.gnome.org's docker version was incompatible with F35
containers for a long time, so we pinned to F34 instead. Now we have
fixed that and updated to F35, but we'll be stuck here for a while
because we won't remember to update again in the future. fedora:latest
seems like a nice compromise between pinning to a particular stable
release vs. using rawhide for CI.