Add GCancellable and GError to the thumbnail API
The methods generate_thumbnail(), save_thumbnail and
create_failed_thumbnail from the class
gnome_desktop_thumbnail_factory offers neither a GError nor a
GCancellable options. For synchronous operations this is not
a true problem, but for asynchronous operations both are
a good idea.

This MR adds those parameters to the method.