Fix idle monitor watch leak
When an idle monitor watch is added and then removed by application,
it is possible that watch->dead is marked when on_watch_added()
is running. However, in this condition libgnome-desktop does not
actually remove this watch in upstream (meta). This idle monitor
watch leak may affect the performance of mutter when it has been
running for a few days.