thumbnail-script: update seccomp filters from flatpak-run.c
It looks like it's been several years since we last updated the
thumbnail sandbox that is copied from Flatpak, and there are several
unresolved sandbox escapes. That's not great.

Hoping we don't forget to update fixes from Flatpak is really not a very
good strategy. Ideally this could would be split into a separate
project. That's been on my to-do list for more than 6 years, though, and
I'm unlikely to get around to it imminently....