gnome-qr: Use the same pre-conditions in the sync and async versions
We have exactly the same requirements, however we were not checking the
async case until the thread was actually spinned, and this might be too
late to figure out what's wrong.

So let's just replicate the same logic in both versions, rather than
returning errors that are not real errors (but rather programmer
errors).